requirements.txt
streamlit
torch==2.0.0
torchvision==0.15.0
Pillow
numpy
faiss-cpu
